refactor(vm): extract PQC precompile contracts
