+++ framework/src/test/java/org/tron/common/runtime/vm/BatchValidateFnDsa512Test.java
@@ -13,8 +13,8 @@ import org.junit.Test;
 import org.tron.common.crypto.pqc.FNDSA512;
 import org.tron.common.crypto.pqc.PQSchemeRegistry;
 import org.tron.common.utils.client.utils.AbiUtil;
+import org.tron.core.vm.PQPrecompiledContracts.BatchValidateFnDsa512;
 import org.tron.core.vm.PrecompiledContracts;
-import org.tron.core.vm.PrecompiledContracts.BatchValidateFnDsa512;
 import org.tron.core.vm.PrecompiledContracts.PrecompiledContract;
 import org.tron.core.vm.config.VMConfig;
 import org.tron.protos.Protocol.PQScheme;
