+++ framework/src/test/java/org/tron/common/runtime/vm/ValidateMultiPQSigTest.java
@@ -27,9 +27,9 @@ import org.tron.common.utils.client.utils.AbiUtil;
 import org.tron.core.capsule.AccountCapsule;
 import org.tron.core.config.args.Args;
 import org.tron.core.store.StoreFactory;
+import org.tron.core.vm.PQPrecompiledContracts.ValidateMultiPQSig;
 import org.tron.core.vm.PrecompiledContracts;
 import org.tron.core.vm.PrecompiledContracts.PrecompiledContract;
-import org.tron.core.vm.PrecompiledContracts.ValidateMultiPQSig;
 import org.tron.core.vm.config.VMConfig;
 import org.tron.core.vm.repository.Repository;
 import org.tron.core.vm.repository.RepositoryImpl;
